Made classes for testings internal to VTableTest
